refactor(entry): standardize import paths
