+++ src/components/EntryList.astro
@@ -1,7 +1,7 @@
 ---
 import { pathList } from '../../constants/path_list';
 import { formatIsoString, formatYYMDString } from '../features/entry/date';
-import type { TinyCollectionEntry } from '../features/entry/tiny_collection_entry';
+import type { TinyCollectionEntry } from '../features/entry/tinyCollectionEntry';
 
 interface Props {
   entries: TinyCollectionEntry[];
