+++ src/features/entry/entryConverter.test.ts
@@ -3,14 +3,14 @@ import path from 'node:path';
 
 import { describe, expect, test } from 'vitest';
 
-import { convertMarkdownToHtml, convertMarkdownToPlainText } from './entry_converter';
+import { convertMarkdownToHtml, convertMarkdownToPlainText } from './entryConverter';
 
 describe('EntryConverter', () => {
   describe('convertMarkdownToHtml', () => {
     test('Markdown text is converted to intended HTML', async () => {
       // Given
-      const markdown = await readFile(path.resolve(__dirname, './fixtures/i-entered-kua.md'), 'utf-8');
-      const expected = await readFile(path.resolve(__dirname, './fixtures/i-entered-kua.html'), 'utf-8');
+      const markdown = await readFile(path.resolve(__dirname, './__fixtures__/i-entered-kua.md'), 'utf-8');
+      const expected = await readFile(path.resolve(__dirname, './__fixtures__/i-entered-kua.html'), 'utf-8');
 
       // When
       const htmlString = await convertMarkdownToHtml(markdown);
@@ -23,8 +23,8 @@ describe('EntryConverter', () => {
   describe('convertMarkdownToPlainText', () => {
     test('Markdown text is converted to intended plain text', async () => {
       // Given
-      const markdown = await readFile(path.resolve(__dirname, './fixtures/i-entered-kua.md'), 'utf-8');
-      const expected = await readFile(path.resolve(__dirname, './fixtures/i-entered-kua.txt'), 'utf-8');
+      const markdown = await readFile(path.resolve(__dirname, './__fixtures__/i-entered-kua.md'), 'utf-8');
+      const expected = await readFile(path.resolve(__dirname, './__fixtures__/i-entered-kua.txt'), 'utf-8');
 
       // When
       const plainText = await convertMarkdownToPlainText(markdown);
