+++ src/features/entry/getSortedEntries.ts
@@ -1,4 +1,4 @@
-import type { TinyCollectionEntry } from './tiny_collection_entry';
+import type { TinyCollectionEntry } from './tinyCollectionEntry';
 
 export function getSortedEntries(entries: TinyCollectionEntry[]): TinyCollectionEntry[] {
   return entries.toSorted((a, b) => (a.data.publishedAt < b.data.publishedAt ? 1 : -1));
