+++ src/pages/index.astro
@@ -4,8 +4,8 @@ import { getCollection } from 'astro:content';
 import { pathList } from '../../constants/path_list';
 import EntryList from '../components/EntryList.astro';
 import Logo from '../components/Logo.astro';
-import { addExcerptToEntries } from '../features/entry/add_excerpt_to_entries';
-import { getSortedEntries } from '../features/entry/get_sorted_entries';
+import { addExcerptToEntries } from '../features/entry/addExcerptToEntries';
+import { getSortedEntries } from '../features/entry/getSortedEntries';
 import { retrieveTranslation } from '../features/locales/i18n';
 import { createHomeStructuredData } from '../features/structured_data/home_structured_data';
 import BaseShell from './_layouts/BaseShell.astro';
